requirements.txt
pyautogui
opencv-python
BeautifulSoup
requests
pillow